docker: don't shadow node_modules
Because we mirror our own filesystem onto `/app`, we also mirror our
empty `node_modules` into the container which means that they have to
be reinstalled locally via `npm i`.

Now installing locally via `npm i` and your own machine's environmentq
is a terrible, terrible idea because we all have different versions of
node, of yarn, of npm, of everything but the kitchen sink, and it
updates the lockfile based on what was resolved in your environment
too.

So don't do that and make sure that `node_modules` is _not_ shadowed.