🚧(frontend) persist token's state
Build a mini PoC of passing match data to the future overlay, that
will open and display match data on cursor's activity.

When a token has the spell-checker styling, it means it is part of an
error or a warning.

Then, I stored for each token its corresponding match, and only logging
it. It will be replaced by a proper overlay mecanism in the next commit.