✨(frontend) open an overlay with misspell or grammar error info
On cursor activity, either clicking or navigating using keyboard,
open an overlay with all relevant information from the spell-checker
server, when encountering a token that has an error.

Language tool doesn't call misspell or grammar issues 'error' but 'match'.

The overlay's HTML is minimal, and doesn't provide any user interaction
(for now), however it lays ground for advanced functionnalities.

When Language Tool suggest too many replacements, I decided to limit the
list of suggestion offered to the user to 5. It's totally arbitrary.