🚸(frontend) position dynamically the spell-checker overlay
Previously, the overlay was always rendered bottom-right position of
the cursor. However, this can lead to undesired behavior, where the cursor
is on the bottom of the page, or the far-right of the editor.

As the overlay has a fixed position, compute its opening position
dynamically, to make sure it will always be usable by the user.