✨(auth) support Agent Connect OIDC provider
This commit introduces a new authentication option within Hedge doc using
Agent Connect, which follows a stricter interpretation of the OIDC specification.

Added environment variables for Agent Connect (AC) client ID and client secret.
AC router is registered conditionally based on the presence of these environment
variables. BUT more environment variables are necessary, please have a look to
the docker compose file.

As AC supports a strict interpretation of the OIDC specification, we had to
implement a custom Passport.js Strategy. New routes were added to handle
the authorize, the authorize callback and the logout callback.

AC supports only the Authorization Code flow. It signs the `/userinfo` response.
It uses email for user lookup, and not the sub as recommended by the spec.
We managed user sessions with Passport.

While authorizing a user, we store its id token in its session data. While login out,
AC requires a parameter named `id_token_hint`, which is the `id_token`.

I relied on Panva library to validate the token and userinfo response. There is
room for improvement on these security checks. An iteration is necessary.

We should absolutely ensure the integrity and the authenticity of the requests and
token payloads.

Enhancement could be done on the error handling and logging strategy.

These changes ensure a robust integration of Agent Connect, facilitating secure,
efficient authentication and logout processes.

In production, users are identified by their email adresses. No data migration should
be required while switching from tchap id to AC.