⚰️(home) remove deprecated "do I have access" button
Removed a link that was dead. It was supposed to explain which users could
access the PAD through the Tchap Identification method. This method is supposed
to whitelist some email domains.

It was broken and outdated according to Julien Dauphant. Plus, as we support
Agent Connect now, it wasn't relevant anymore.