♻️(frontend) enhance Agent Connect logout function
If no ID token is sotred in session's data, the logout function early returns.
If the code early returns, sotring a state in session data is useless.
The state is passed while calling the OIDC provider to logout.

I noticed it while re-reading my code, mybad, it's a slighlty improved code.