🩹(frontend) call the right logout route when using Agent Connect
I haven't noticed yet, but my view template wasn't calling the right
logout route when signed-in using AgentConnect.

AgentConnect requires a logout callback view to properly logout the users.
It's a different implementation from the already supported Oauth2 providers,
which all share a common route /logout.

It now routes conditionnally on the appropriate logout view, based on which
Auth provider is enabled. Please note that my code is not bullet proof.

We decided to support only one OAuth2 provider at a time, although the code does
not enforce a limit on the number of OAuth2 providers. This means that if both
AgentConnect and another provider are activated simultaneously, the second
provider would use the logout route intended for AgentConnect. Since no ID token
would be stored in the session, the user would be redirected back to the /logout view.
Despite this, the logout process would still function correctly.