✨(backend) support prompt=None Authentication Request with Agent Connect
This parametrization of the Authentication Request is necessary for the True
SSO experience. As a La Suite service, we would like to determine if the user
has an active session with Agent Connect, if not its login is required.

The OIDC provider should redirect to our backend, so the user agent (e.g Chrome)
is not blocked on the Agent Connect loging screen.

Next commit add the logic to call the "silent authentication" and handle
the rendering of the homepage if the login is required.