🚸(backend) enable true SSO using Agent Connect
Logged-in silently the user if her has an active session with the OIDC
provider, namely Agent Connect.

If the user cannot be logged-in, the default homepage is shown. It fits
design and UX requirements.

Btw, not a huge fan of manual URL building, but that fits codebase style
(plus, it's straightforward and simple to grasp).