🚧(frontend) introduce tutorial presenting spell-checker
Using Intro.js, bootstrap a basic tooltip to inform users about the newly
introduced spell-checker feature in beta.

The tutorial is essential to ensure users are aware of this new functionality
and how to enable or disable it. I've added a 'boilerplate' content, open to
suggestions and improvements.

Will be enhanced in upcoming commits.