🚸(frontend) show tutorial only once
I struggled using the IntroJs API and might have missed something.
There's no built-in way to force the tutorial to show only once without
displaying a checkbox to the user.

I prefer not to give users the option to skip the tutorial. It should be
shown only once.

Without using the 'dontShowAgain' option, the 'isActive' method always
returns true, even if the cookie is set to true.

To handle this, I used the 'dontShowAgain' option, set the cookie
value, and override some styling to hide the checkbox.

While developing, clear the a cookie named 'introjs-dontShowAgain'.